# src/main/resources/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Web VolunteerDTO Coordination System - OpenAPI 3.0
  description: |-
    This is a Web Volunteer Coordination System based on the OpenAPI 3.0 specification. 
  contact:
    email: 260292@student.pwr.edu.pl
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: localhost:8080/
...
paths:
...
  /tasks:
    get:
      tags:
        - task
      summary: Retrieve all tasks
      operationId: getAllTasks
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskDTO'

    post:
      tags:
        - task
      summary: Create a new task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskDTO'
      responses:
        '201':
          description: TaskDTO created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'
  /tasks/{taskId}:
    get:
      tags:
        - task
      summary: Retrieve detailed task info
      operationId: getTaskSkills
      parameters:
        - name: taskId
          in: path
          required: true
          description: ID of the task to list
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskWithSkillsDTO'
    put:
      tags:
        - task
      summary: Update a task
      operationId: updateTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: ID of the task to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskDTO'
      responses:
        '200':
          description: TaskDTO updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'

    delete:
      tags:
        - task
      summary: Delete a task
      operationId: deleteTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: ID of the task to delete
          schema:
            type: integer
      responses:
        '204':
          description: TaskDTO deleted successfully
        '404':
          description: TaskDTO not found
        '500':
          description: Internal server error
...
  /organizer:
    get:
      tags:
        - organizer
      summary: Retrieve all organizers
      operationId: getAllOrganizers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizerDTO'

    post:
      tags:
        - organizer
      summary: Create a new organizer
      operationId: createOrganizer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizerDTO'
      responses:
        '201':
          description: OrganizerDTO created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerDTO'
  /organizer/{organizerId}:
    get:
      tags:
        - organizer
      summary: Retrieve detailed organizer info by ID
      operationId: getOrganizerById
      parameters:
        - name: organizerId
          in: path
          required: true
          description: ID of the organizer to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerWithUserDTO'

    put:
      tags:
        - organizer
      summary: Update an organizer
      operationId: updateOrganizer
      parameters:
        - name: organizerId
          in: path
          required: true
          description: ID of the organizer to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizerDTO'
      responses:
        '200':
          description: OrganizerDTO updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerDTO'

    delete:
      tags:
        - organizer
      summary: Delete an organizer
      operationId: deleteOrganizer
      parameters:
        - name: organizerId
          in: path
          required: true
          description: ID of the organizer to delete
          schema:
            type: integer
      responses:
        '204':
          description: OrganizerDTO deleted successfully
        '404':
          description: OrganizerDTO not found
        '500':
          description: Internal server error
  /skill:
    get:
      tags:
        - skill
      summary: Retrieve all skills
      operationId: getAllSkills
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SkillDTO'

    post:
      tags:
        - skill
      summary: Create a new skill
      operationId: createSkill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillDTO'
      responses:
        '201':
          description: SkillDTO created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDTO'
        '400':
          description: SkillDTO already exists
  /skill/{skillId}:
    delete:
      tags:
        - skill
      summary: Delete a skill
      operationId: deleteSkill
      parameters:
        - name: skillId
          in: path
          required: true
          description: ID of the skill to delete
          schema:
            type: integer
      responses:
        '204':
          description: SkillDTO deleted successfully
        '404':
          description: SkillDTO not found
        '500':
          description: Internal server error
...
components:
  schemas:
...
    TaskDTO:
      type: object
      required:
        - status
        - name
        - eventId
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
        eventId:
          type: integer
          format: int64
          example: 198772
        assigneeId:
          type: integer
          format: int64
          example: 198772
        deadline:
          type: string
          format: date-time
        xLocation:
          type: number
        yLocation:
          type: number
        description:
          type: string
        status:
          type: string
          description: TaskDTO CrewStatus
          example: to do
          enum:
            - to do
            - in progress
            - done
      xml:
        name: task
    TaskWithSkillsDTO:
      type: object
      required:
        - status
        - name
        - eventId
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
        eventId:
          type: integer
          format: int64
          example: 198772
        assigneeId:
          type: integer
          format: int64
          example: 198772
        deadline:
          type: string
          format: date-time
        xLocation:
          type: number
        yLocation:
          type: number
        description:
          type: string
        status:
          type: string
          description: TaskDTO CrewStatus
          example: to do
          enum:
            - to do
            - in progress
            - done
        skills:
          type: array
          items:
            $ref: '#/components/schemas/SkillDTO'
      xml:
        name: taskWithSkills
...
    OrganizerDTO:
      type: object
      required:
        - brandName
        - userId
      properties:
        id:
          type: integer
          format: int64
          example: 100000
        brandName:
          type: string
          example: BigCompany
        userId:
          type: integer
          format: int64
          example: 100000
      xml:
        name: organizer
    OrganizerWithUserDTO:
      type: object
      required:
        - brandName
        - userId
      properties:
        id:
          type: integer
          format: int64
          example: 100000
        brandName:
          type: string
          example: BigCompany
        user:
          $ref: '#/components/schemas/UserDTO'
      xml:
        name: organizerWithUser
    UserDTO:
      type: object
      required:
        - name
        - email
        - password
        - phoneNumber
      properties:
        id:
          type: integer
          format: int64
          example: 10
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
        phoneNumber:
          type: string
          example: '12345'
      xml:
        name: user
...
    SkillDTO:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: Ticketing
      xml:
        name: skill
...